server: Export more container metadata for VM containers
VM base container runtimes (e.g. Clear Containers) will run each pod
in a VM and will create containers within that pod VM. Unfortunately
those runtimes will get called by ocid with the same commands
(create and start) for both the pause containers and subsequent
containers to be added to the pod namespace. Unless they work around
that by e.g. infering that a container which rootfs is under
"/pause" would represent a pod, they have no way to decide if they
need to create/start a VM or if they need to add a container to an
already running VM pod.

This patch tries to formalize this difference through pod
annotations. When starting a pause container/pod, we now add 2
annotations for the container type (Infrastructure or not) and the
sandbox id. This will allow VM based container runtimes to handle 2
things:

- Decide if they need to create a pod VM or not
- Keep track of which pod ID runs in a given VM, so that they
  know to which pod they have to add non infrastructure containers.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>